Refactor matchKeywordInSubheadings
